requirements.txt
pytest==6.2.5
pytest-mock==3.6.1
pytest-pythonpath==0.7.3
pytest-html==3.1.1
pytest-cov==3.0.0
pytest-django==4.4.0